Refactor files name
