+++ src/containers/courses/Course.js
@@ -1,0 +1,39 @@
+import React, { Component } from 'react';
+import { connect } from "react-redux";
+import Table from '../../components/utils/table'
+import * as courseAction from "../../actions/course-action";
+import './style/course.css';
+
+class Courses extends Component {
+
+  headers = {
+    name: 'Course',
+    users: 'Users',
+    owner: 'Owner',
+    date: 'Created at',
+    enrolled: 'Enrolled',
+    actions: 'Actions' // check how we're going to deal with solving
+  };
+
+  componentDidMount() {
+    this.props.onInitCourses()
+  }
+
+  // renderContent() {
+  //   if (this.props.isLoading) {
+  //     return (<div className="spinner" />)
+  //   }
+
+  //   return (<Table headers={this.headers} data={this.props.courses} />)
+  // };
+
+  render() {
+    const table = (this.props.isLoading) ? <div className="spinner" /> 
+                  : <Table headers={this.headers} data={this.props.courses} />
+    return (
+      <div className='course-container'>
+        {table}
+      </div>
+    );
+  }
+}
